moving add all intent handler to separate file
