+++ lambda/custom/handlers/addAllToChannel.js
@@ -1,0 +1,37 @@
+const { ri } = require('@jargon/alexa-skill-sdk');
+const { replaceWhitespacesFunc, login, getRoomId, addAll } = require('../helperFunctions');
+
+const AddAllToChannelIntentHandler = {
+	canHandle(handlerInput) {
+		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
+			handlerInput.requestEnvelope.request.intent.name === 'AddAllToChannelIntent';
+	},
+	async handle(handlerInput) {
+		try {
+
+			const {
+				accessToken,
+			} = handlerInput.requestEnvelope.context.System.user;
+			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.addallchannelname.value;
+			const channelName = replaceWhitespacesFunc(channelNameData);
+
+			const headers = await login(accessToken);
+			const roomid = await getRoomId(channelName, headers);
+			const speechText = await addAll(channelName, roomid, headers);
+			const repromptText = ri('GENERIC_REPROMPT');
+
+			return handlerInput.jrb
+				.speak(speechText)
+				.speak(repromptText)
+				.reprompt(repromptText)
+				.withSimpleCard(ri('ADD_ALL_TO_CHANNEL.CARD_TITLE'), speechText)
+				.getResponse();
+		} catch (error) {
+			console.error(error);
+		}
+	},
+};
+
+module.exports = {
+	AddAllToChannelIntentHandler,
+};
